+++ apps/web/src/domain/challenge/my-challenge/mission-calendar/ui/MissionNotStartedIcon.tsx
