+++ apps/web/src/domain/challenge/my-challenge/section/mission-submit/MissionSubmitRegularSection.tsx
@@ -2,7 +2,7 @@ import { useParams } from 'next/navigation';
 import BonusMissionModal from '../../../bonus-mission/modal/BonusMissionModal';
 import DashboardCreateReviewModal from '../../../dashboard/modal/DashboardCreateReviewModal';
 import MobileReviewModal from '../../modal/MobileReviewModal';
-import MissionSubmitButton from '../../mission/MissionSubmitButton';
+import MissionSubmitButton from '../../mission/ui/MissionSubmitButton';
 import MissionToast from '../../mission/MissionToast';
 import LinkInputSection from '../LinkInputSection';
 import { MissionSubmitListForm } from '../mission-submit-list-form';
